Add count page component
add count page component for spilt count page into index
