+++ src/index.jsx
@@ -1,32 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDom from 'react-dom';
-import CountButton from './countButton';
-import CountNumberButton from './countNumberButton';
+import CountPage from './countPage';
 
 function App() {
-  const [count, setCount] = useState(0);
-
-  function handleClick() {
-    setCount((prev) => prev + 1);
-  }
-
-  function handleClickNumber(value) {
-    setCount((prev) => prev + value);
-  }
   return (
-    <div>
-      <p>
-        Counter
-      </p>
-      <p>
-        <CountButton count={count} onClick={handleClick} />
-      </p>
-      <p>
-        {[1, 2, 3, 4, 5].map((i) => (
-          <CountNumberButton value={i} onClick={() => handleClickNumber(i)} />
-        ))}
-      </p>
-    </div>
+    <CountPage />
   );
 }
